pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-analysis"
version = "1.0.4"
description = "Extracts keywords from text, runs sentiment analysis on those keywords and publish to redis"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
dependencies = [
  "numpy",
  "pyyaml",
  "python-dotenv",
  "redis",
  "pandas",
  "spacy",
  "keybert",
  "transformers",
  "torch",
  "scikit-learn",
  "nltk",
  "data-models @ git+https://github.com/zeidlitz/data-models.git"
]

[project.scripts]
run-analysis = "data_analysis.main:main"
